backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-dbbackup>=4.0.0
django-silk>=5.0.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
python-dotenv==1.2.1
easy-thumbnails==2.10.1
Faker==40.5.1
huey>=2.5.0
pillow==12.1.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
freezegun==1.5.5
coverage==7.13.4
factory-boy==3.3.3
redis>=4.0.0
ruff==0.15.2
requests==2.32.5
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0

pydub>=0.25.1

# Production server
gunicorn>=23.0,<24.0
mysqlclient>=2.2,<3.0
